% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGatingSetFromFaust.R
\name{makeGatingSetFromFAUST}
\alias{makeGatingSetFromFAUST}
\title{Make a Gated Gating Set From a FAUST Result}
\usage{
makeGatingSetFromFAUST(gatingSet = NULL, faustProjectDirectory = NULL)
}
\arguments{
\item{gatingSet}{\code{GatingSet} object.}

\item{faustProjectDirectory}{\code{character} path to a FAUST project directory containing FAUST results.}
}
\value{
\code{NULL}
}
\description{
For each cell population discovered by FAUST, we gate the cell population according to the
FAUST marker order and add it to the GatingSet.
}
\examples{
\dontrun{
library(flowWorkspace)
library(ggcyto)
library(dplyr)
library(tidyr)
gs <- load_gs("/Users/gfinak/phenogs")
makeGatingSetFromFAUST(faustProjectDirectory =
"/shared/silo_researcher/gottardo_r/gfinak_working/CITN07/FAUST_pheno_full_remapped_persample_consistentPars",
gatingSet = gs[1:2])
 recompute(gs[1:2])
stats <- getPopStats(gs[1:2], path = "full")
}
}
