% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGatePath.R
\name{plotGatePath}
\alias{plotGatePath}
\title{Plot all the gates along a path.}
\usage{
plotGatePath(gh, path, stop = "root", ...)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{path}{a character string representing a node path}

\item{stop}{a string representing the ancestor node where the plot stops.All the gates prior to that node (and include that node) will be skipped.
Default is "root", which means trace all the way back to the first node in the path}

\item{...}{arguments passed to ggcyto::autoplot}
}
\description{
Plots the gates along a path from the first node in the path to the terminal node.
This is like ploting the gating scheme for a specific population. The \code{stop} argument is a node or path that specifies where to  stop plotting upstream. For example, if the path is "A/B/C", and stop is "A", then gates "B" and "C" will be plotted.
It is recommended to pass \code{bins=256} to the plotting routine to get sufficient resolution.
}
\examples{
\dontrun{
dataDir <- system.file("extdata",package="flowWorkspaceData")
library(flowWorkspace)
gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE))
library(ggcyto)
autoplot(gs[[1]], "singlets/CD3+/CD4") #plot only the leaf gate
#plot all the three gates in the gating path
plotGatePath(gs[[1]], "singlets/CD3+/CD4", strip.text = "gate")
plotGatePath(gs[[1]], "singlets/CD3+/CD4/CCR7- 45RA-", stop = "CD3+")#exclude the first two gates
plotGatePath(gs[[1]], "/not debris/singlets", strip.text = "gate")
}
}
